Authentication/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
bcrypt==3.2.0
mysql-connector-python==8.0.25
